refactor(core): split sign in experience lib
