+++ packages/core/src/lib/sign-in-experience/index.ts
@@ -1,0 +1,20 @@
+import { Branding, BrandingStyle, TermsOfUse } from '@logto/schemas';
+
+import assertThat from '@/utils/assert-that';
+
+export * from './sign-in-methods';
+
+export const validateBranding = (branding: Branding) => {
+  if (branding.style === BrandingStyle.Logo_Slogan) {
+    assertThat(branding.slogan?.trim(), 'sign_in_experiences.empty_slogan');
+  }
+
+  assertThat(branding.logoUrl.trim(), 'sign_in_experiences.empty_logo');
+};
+
+export const validateTermsOfUse = (termsOfUse: TermsOfUse) => {
+  assertThat(
+    !termsOfUse.enabled || termsOfUse.contentUrl,
+    'sign_in_experiences.empty_content_url_of_terms_of_use'
+  );
+};
