+++ packages/core/src/lib/sign-in-experience/sign-in-methods.test.ts
@@ -1,73 +1,15 @@
-import { BrandingStyle, SignInMethodState, ConnectorType } from '@logto/schemas';
+import { SignInMethodState, ConnectorType } from '@logto/schemas';
 
 import {
   mockAliyunDmConnector,
   mockFacebookConnector,
   mockGithubConnector,
-  mockBranding,
   mockSignInMethods,
 } from '@/__mocks__';
 import RequestError from '@/errors/RequestError';
-import {
-  isEnabled,
-  validateBranding,
-  validateSignInMethods,
-  validateTermsOfUse,
-} from '@/lib/sign-in-experience';
+import { isEnabled, validateSignInMethods } from '@/lib/sign-in-experience';
 
 const enabledConnectors = [mockFacebookConnector, mockGithubConnector];
-
-describe('validate branding', () => {
-  test('should throw when the UI style contains the slogan and slogan is empty', () => {
-    expect(() => {
-      validateBranding({
-        ...mockBranding,
-        style: BrandingStyle.Logo_Slogan,
-        slogan: '',
-      });
-    }).toMatchError(new RequestError('sign_in_experiences.empty_slogan'));
-  });
-
-  test('should throw when the logo is empty', () => {
-    expect(() => {
-      validateBranding({
-        ...mockBranding,
-        style: BrandingStyle.Logo,
-        logoUrl: ' ',
-        slogan: '',
-      });
-    }).toMatchError(new RequestError('sign_in_experiences.empty_logo'));
-  });
-
-  test('should throw when the UI style contains the slogan and slogan is blank', () => {
-    expect(() => {
-      validateBranding({
-        ...mockBranding,
-        style: BrandingStyle.Logo_Slogan,
-        slogan: ' \t\n',
-      });
-    }).toMatchError(new RequestError('sign_in_experiences.empty_slogan'));
-  });
-
-  test('should not throw when the UI style does not contain the slogan and slogan is empty', () => {
-    expect(() => {
-      validateBranding({
-        ...mockBranding,
-        style: BrandingStyle.Logo,
-      });
-    }).not.toThrow();
-  });
-});
-
-describe('validate terms of use', () => {
-  test('should throw when terms of use is enabled and content URL is empty', () => {
-    expect(() => {
-      validateTermsOfUse({
-        enabled: true,
-      });
-    }).toMatchError(new RequestError('sign_in_experiences.empty_content_url_of_terms_of_use'));
-  });
-});
 
 describe('check whether the social sign-in method state is enabled', () => {
   it('should be truthy when sign-in method state is primary', () => {
